Reorganize files.  Support loading KHARMA dumps with equivalent language to iharm3d
