+++ docs/ref_dumps.rst
@@ -8,7 +8,7 @@ Reading HARM HDF5 Output
 
 These are functions for reading files of the form documented on the AFD docs `wiki <https://github.com/AFD-Illinois/docs/wiki/Output-Format>`_.  This is the format used by modern iterations of ``iharm3d`` and ``ebhlight``, as well as the HARM implementation in this package.  Several tools for converting output from other codes into this format can be found at `EHT-babel <https://github.com/AFD-Illinois/EHT-babel>`_.
 
-There are two ways to read HARM HDF5 output with the pyHARM tools: the high level interface, :mod:`pyHARM.ana.iharm_dump`, which allows accessing derived variables of several types in a ``dict``-like interface, or :mod:`pyHARM.h5io`, which reads and writes arrays of the primitive variables used by HARM.
+There are two ways to read HARM HDF5 output with the pyHARM tools: the high level interface, :mod:`pyHARM.ana.iharm_dump`, which allows accessing derived variables of several types in a ``dict``-like interface, or :mod:`pyHARM.io`, which reads and writes arrays of the primitive variables used by HARM.
 
 IharmDump interface
 -------------------
@@ -19,5 +19,5 @@ IharmDump interface
 h5io interface
 --------------
 
-.. automodule:: pyHARM.h5io
+.. automodule:: pyHARM.io
    :members:
