+++ pyHARM/grmhd/phys.py
