+++ pyHARM/util.py
